feat: extract KnowledgeSearchService from KnowledgeStoreService
Move search(), searchReflections(), BM25/FTS5 query construction,
RRF hybrid merge logic, and snippet extraction into a dedicated
KnowledgeSearchService (323 lines). KnowledgeStoreService delegates
via composition, preserving the existing facade API.
